+++ client/src/extension.ts
@@ -141,6 +141,7 @@ export async function activate(ctx: vscode.ExtensionContext) {
   ctx.subscriptions.push(statusBarEnvItem);
 
   ctx.subscriptions.push(vscode.commands.registerCommand('iiq-dev-accelerator.importFile', () => iiqCommands.importFile()));
+  ctx.subscriptions.push(vscode.commands.registerCommand('iiq-dev-accelerator.importFileFromExplorer', iiqCommands.importFileFromExplorer, iiqCommands));
   ctx.subscriptions.push(vscode.commands.registerCommand('iiq-dev-accelerator.runTask', () => iiqCommands.runTask()));
   ctx.subscriptions.push(vscode.commands.registerCommand('iiq-dev-accelerator.runTaskWithAttr', () => iiqCommands.runTaskWithAttr()));
   ctx.subscriptions.push(vscode.commands.registerCommand('iiq-dev-accelerator.runRule', () => iiqCommands.runRule()));